feat: :sparkles: Created the factory, and also adjuste the kpis app
