+++ src/express/Express.ts
@@ -1,7 +1,6 @@
 import express, { Application } from "express";
 import ExpressEnvironment from "./config/ExpressEnvironment";
 import ExpressRouter from "./routes/ExpressRouter";
-import path from "path";
 
 export default class Express {
   private readonly app: Application;
@@ -21,7 +20,6 @@ export default class Express {
   config = (): void => {
     this.app.use(express.json());
     this.app.use(express.urlencoded({ extended: true }));
-    this.app.use(express.static(path.join(__dirname, "../frontend")));
   };
 
   routes = (): void => {
